requirements.txt
streamlit==1.22.0
openai==0.27.0
requests==2.28.2
google-auth==2.15.0
google-api-python-client==2.80.0
pandas==2.0.3
python-dotenv==1.0.0
